Need to start converting from AlgorithmUser class.

+++ CudaMONSEL/CudaMONSEL/gov/nist/microanalysis/EPQLibrary/CzyzewskiMottCrossSection.cuh
@@ -27,7 +27,7 @@ namespace CzyzewskiMottCrossSection
       void loadTables(int atomicNo);
 
       const ElementT& mElement;
-      MatrixXf mValues;
+      MatrixXd mValues;
    };
 
    double getSpecialEnergy(int index);
